+++ tests/core/calculators/vasp/test_vasp.py
@@ -12,7 +12,7 @@ from ase.atoms import Atoms
 from ase.build import bulk
 from ase.calculators.singlepoint import SinglePointDFTCalculator
 from ase.calculators.vasp import Vasp as Vasp_
-from ase.constraints import FixAtoms, FixBondLength
+from ase.constraints import FixAtoms
 from ase.io import read
 from pymatgen.io.vasp.sets import MPRelaxSet, MPScanRelaxSet
 
@@ -94,7 +94,7 @@ def test_presets2():
 def test_presets_mp():
     atoms = bulk("Co") * (2, 2, 1)
     atoms[-1].symbol = "Fe"
-    parameters = MPtoASEConverter(atoms=atoms).convert_dict_set(MPScanRelaxSet())
+    parameters = MPtoASEConverter(atoms=atoms).convert_input_set(MPScanRelaxSet())
     calc = Vasp(atoms, xc="scan", **parameters)
     assert calc.xc.lower() == "scan"
     assert calc.parameters["algo"].lower() == "all"
@@ -736,7 +736,7 @@ def test_magmoms(atoms_mag, atoms_nomag, atoms_nospin):
 
     atoms = bulk("Cu") * (2, 2, 1)
     atoms[-1].symbol = "Fe"
-    parameters = MPtoASEConverter(atoms=atoms).convert_dict_set(MPScanRelaxSet())
+    parameters = MPtoASEConverter(atoms=atoms).convert_input_set(MPScanRelaxSet())
     calc = Vasp(atoms, **parameters)
     atoms.calc = calc
     assert atoms.get_chemical_symbols() == ["Cu", "Cu", "Cu", "Fe"]
@@ -745,7 +745,7 @@ def test_magmoms(atoms_mag, atoms_nomag, atoms_nospin):
     atoms = bulk("Cu") * (2, 2, 1)
     atoms[-1].symbol = "Fe"
     atoms.set_initial_magnetic_moments([3.14] * (len(atoms) - 1) + [1.0])
-    parameters = MPtoASEConverter(atoms=atoms).convert_dict_set(MPScanRelaxSet())
+    parameters = MPtoASEConverter(atoms=atoms).convert_input_set(MPScanRelaxSet())
     calc = Vasp(atoms, **parameters)
     atoms.calc = calc
     assert atoms.get_initial_magnetic_moments().tolist() == [3.14] * (
@@ -1265,7 +1265,7 @@ def test_setups():
     assert calc.parameters["setups"]["Cu"] == ""
 
     atoms = bulk("Cu")
-    parameters = MPtoASEConverter(atoms=atoms).convert_dict_set(MPScanRelaxSet())
+    parameters = MPtoASEConverter(atoms=atoms).convert_input_set(MPScanRelaxSet())
     calc = Vasp(atoms, **parameters)
     assert calc.parameters["setups"]["Cu"] == "_pv"
 
@@ -1359,13 +1359,6 @@ def test_constraints():
     atoms.calc = calc
     assert isinstance(atoms.constraints[0], FixAtoms)
 
-    atoms = bulk("Cu") * (2, 1, 1)
-    atoms.set_constraint(FixBondLength(0, 1))
-    with pytest.raises(
-        ValueError, match="Atoms object has a constraint that is not compatible"
-    ):
-        calc = Vasp(atoms)
-
 
 def test_envvars():
     with change_settings(
@@ -1421,7 +1414,7 @@ def test_bad_pmg_converter():
 
 def test_pmg_input_set():
     atoms = bulk("Cu")
-    parameters = MPtoASEConverter(atoms=atoms).convert_dict_set(MPRelaxSet())
+    parameters = MPtoASEConverter(atoms=atoms).convert_input_set(MPRelaxSet())
     calc = Vasp(atoms, incar_copilot="off", **parameters)
     assert calc.parameters == {
         "algo": "fast",
@@ -1451,7 +1444,7 @@ def test_pmg_input_set():
 def test_pmg_input_set2():
     atoms = bulk("Fe") * (2, 1, 1)
     atoms[0].symbol = "O"
-    parameters = MPtoASEConverter(atoms=atoms).convert_dict_set(MPRelaxSet())
+    parameters = MPtoASEConverter(atoms=atoms).convert_input_set(MPRelaxSet())
     calc = Vasp(atoms, incar_copilot="off", **parameters)
     assert calc.parameters == {
         "algo": "fast",
@@ -1490,7 +1483,7 @@ def test_ldau_mp():
     atoms.pbc = True
     atoms[0].position += 0.1
     atoms[1].position -= 0.1
-    parameters = MPtoASEConverter(atoms=atoms).convert_dict_set(MPRelaxSet())
+    parameters = MPtoASEConverter(atoms=atoms).convert_input_set(MPRelaxSet())
     assert len(parameters["ldauu"]) == 3
     assert parameters["ldauu"] == [0, 0, 3.9]
     assert len(parameters["magmom"]) == 4