+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project generally adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.6]
+
+### Changed
+
+- Changed the default `SYMPREC` in the VASP `relax_job` to 1e-6 instead of 1e-8
+
+## [1.0.5]
+
+### Fixed
+- Fixed bug when trying to use Custodian to run a single-point VASP calculation on an Atoms object with constraints
+- Fixed various type hints
+
 ## [1.0.4]
 
 ### Added